improved code structure of WorldObjects
WorldObject is the base class of anything on the world, like the player or the cubes (for now).

WorldObjects contain the logic data, like where the object is, and the Models class contains the data of how the objects look and updates it based on how the respective WorldObject changes.